+++ rest-openehr/src/main/java/org/ehrbase/rest/openehr/OpenehrDirectoryController.java
@@ -140,13 +140,17 @@ public class OpenehrDirectoryController extends BaseController {
         }
 
         // Insert New folder
-        ObjectVersionId folderId = this.folderService.create(
+        Optional<FolderDto> dtoOptional = this.folderService.create(
                 ehrId,
-                folder,
-                null,
-                null,
-                null
-        );  // TODO-526: add default committer and description handling
+                folder
+        );
+        ObjectVersionId folderId;
+        var folderDto = dtoOptional.orElseThrow(() -> new InternalServerException("Error creating folder."));
+        if (folderDto.getUid() instanceof ObjectVersionId) {
+            folderId = (ObjectVersionId) folderDto.getUid();
+        } else {
+            throw new InternalServerException("Unexpected folder DTO ID type.");
+        }
 
         // Fetch inserted folder for response data
         Optional<FolderDto> newFolder = this.folderService.get(folderId, null);
@@ -353,13 +357,9 @@ public class OpenehrDirectoryController extends BaseController {
 
         // Update folder and get new version
         Optional<FolderDto> updatedFolder = this.folderService.update(
-                folderId,
-                folder,
-                ehrId,
-                null,
-                null,   // TODO-526: add default committer handling
-                null,   // TODO-526: placeholder for now
-                null    // TODO-526: placeholder for now
+            ehrId,
+            folderId,
+            folder
         );
 
 
@@ -419,8 +419,7 @@ public class OpenehrDirectoryController extends BaseController {
         // Check version conflicts and throw precondition failed exception if not
         checkDirectoryVersionConflicts(folderId, ehrId);
 
-        // TODO-526: add default committer and description handling
-        this.folderService.delete(folderId, null, null, null, null);
+        this.folderService.delete(ehrId, folderId);
         this.ehrService.removeDirectory(ehrId);
         return createDirectoryResponse(HttpMethod.DELETE, null, accept, null, ehrId);
     }