+++ CHANGELOG.md
@@ -5,8 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## [Unreleased]
+
+### Added
+
+### Changed
+
+### Fixed
+
+## [0.17.1] (beta)
+
+### Added
+- Default handling for audit metadata (see: https://github.com/ehrbase/ehrbase/pull/552)
+
+### Changed
+- Updated the SDK dependency to the latest version (see: https://github.com/ehrbase/ehrbase/pull/565)
+- Refactored versioned object (interfaces) on service and access layer (see: https://github.com/ehrbase/ehrbase/pull/552)
+
+### Fixed
+- Assigner in DV_IDENTIFIER not selected in aql (see: https://github.com/ehrbase/ehrbase/pull/561)
+- ehr_status.uuid not selects via aql (see: https://github.com/ehrbase/ehrbase/pull/561)
+- DB migration file conflict (see: https://github.com/ehrbase/ehrbase/pull/564)
+- Ddmin delete of multiple status versions (see: https://github.com/ehrbase/ehrbase/pull/552)
+
+## [0.17.0] (beta)
 
 ### Added
 - Implement validation of compositions using external FHIR TS (see: https://github.com/ehrbase/ehrbase/pull/493)
@@ -18,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Folder handling (update, delete and missing audits) (see: https://github.com/ehrbase/ehrbase/pull/529)
+- Fixed and refactored handling of audits and versioned objects (see: https://github.com/ehrbase/ehrbase/pull/552/)
 
 ## [0.16.0] (beta)
 
@@ -234,7 +257,9 @@ Starting with the next release this file will provide a proper overview.
 - Fixed error regarding DISTINCT operator in AQL (https://github.com/ehrbase/ehrbase/issues/50)
 - Fixed null pointer exceptions that could occur in persistent compositions
 
-[unreleased]: https://github.com/ehrbase/ehrbase/compare/v0.16.0...HEAD
+[unreleased]: https://github.com/ehrbase/ehrbase/compare/v0.17.1...HEAD
+[0.17.1]: https://github.com/ehrbase/ehrbase/compare/v0.17.0...v0.17.1
+[0.17.0]: https://github.com/ehrbase/ehrbase/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/ehrbase/ehrbase/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/ehrbase/ehrbase/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/ehrbase/ehrbase/compare/v0.13.0...v0.14.0